Add more variable checking for Procera issue
For years now we've been seeing application and protocol names being
corrupted. These changes are another attempt to narrow down what is
going on. First, in a previous attempt to win this battle I moved our
variables from the stack to malloc'ed memory. It didn't make any
difference, so I went back to using stack variables to get rid of the
dynamic memory allocation overhead.

Now, according to the NAVL documentation, the navl_proto_get_name()
function is supposed to return NULL and not touch the contents of the
buffer we pass if there is an error. Since we effectively set it to
empty with buffer[0] = 0 we don't care if it returns NULL since we rely
on the buffer staying untouched. If the documentation is wrong and the
buffer does get changed, that could cause the issues we've been seeing.
So I added new logic to check the return value each time we call that
function. I also added counters so we can see how often it returns NULL
when called to get the name of the classified application, and when
called to get any of the names while iterating the protochain. I also
supplemented the vineyard garbage counter to keep track of the amount
of garbage we see in both of those cases.
@changelog
[NGFW-9482]