Fix the way we get protocol names
I was using isalnum() to detect garbage in protocol names returned
by the NAVL library. Turns out some of the names include a period so
were were incorrectly counting junk, making it harder to see the actual
problem of garbage characters. I changed to isascii() which is what the
Java code looks for. I also went back to our original method of
creating an array or protocol names on startup, rather than calling
navl_proto_get_name repeatedly. This avoids the overhead of extra
function calls, and brings back the ability to track how many packets
we see for each protocol, which is kinda nifty. I also switch from
snprintf to strcat when building the protochain since it should
be more efficient.


Former-commit-id: bf5ca7af13b53a2b12d7531e86f6f182bdc986bd