Fix handling of application, protochain, and detail strings
For a long time, we have stored strings in the session object by using
malloc and memcpy for the new string, and then doing a pointer swap to
make the new string active, followed by a free of the old string. A
recent valgrind run to track down the source of the garbage characters
we see on occasion generated a hit on freed memory being read. While
I expected our original approach to be thread safe with the way I
implemented the classify and network thread message queues, clearly
there are situations where my original method was failing.

I changed the logic to allocate a pair of buffers for each of the
strings we store when the session object is created, and we keep an
index for each to select which of the two is active at any time. During
update, we copy the new string to the inactive buffer, and then toggle
the active index with (value ^= 1). So now if a thread is still trying
to read the previously active string after an update, the data will
still be valid and available. This also has the side benefit of taking
the memory allocation calls out of the update function, since all of
the buffer pairs are allocated when the session object is created.

I also fixed the protochain strncat calls to pass the buffer size
minus one since strncat can touch n+1 bytes.

NGFW-9482 #code-completed


Former-commit-id: a4769ad155eae5a4f4b9c4349fd66f87cfe83eb1