Fix bdamserver package
Lots of fixes and cleanup. Removed the old spamassassin plugin stuff.
Moved the prerm cleanup to postrm since that's where it belongs based on
the dpkg flow chart, and it now actually does the cleanup. Created a
user for the bdamserver, and put all of the file and directory creation
and user and chown stuff in the postinst with corresponding cleanup in
the postrm. Added logrotate script. Enabled the LogAppend option.
Added WorkingDirectory to service definition so the daemon can
find the update library.

@changelog (for auto generation or remove if not wanted)