Fix memory leak in session cleanup
Fixed a race condition in the session REMOVE logic. The netclient
thread was deleting the object from the session table before the
classify thread could process the remove message, so we never cleaned
up the NAVL connection object. The classify thread now handles the
session delete for TCP and UDP sessions from NGFW just like in previous
versions, and the netclient thread handles the delete for IP4 and IP6
sessions from MFW.

@changelog