Adjust warning messages
I added a conditional to the session not found warning messages so they
will not be generated when running with the -N runtime argument that we
use to adjust the log output when run from the packet daemon. Because
of the way we handle traffic in MFW we know it's possible for us to
send daemon commands for sessions after they have been cleaned up.
Since the behavior is expected, we don't really need the warning
messages filling up the log.