Rework to improve MFW interoperability
The daemon now has a -mfw runtime argument for use by the packet daemon.
When this flag is set, we do not create the classify thread, message
queue, or session table. The NAVL library init and shutdown is handled
directly by the network thread, with all raw packets being passed
to navl_classify directly. This gives us the correct classification
result on the first packet, and eliminates the session table and
all related complications, including the need for session CREATE and
REMOVE commands which were only needed to trigger the navl_conn_create
and navl_conn_destroy which aren't used in raw IP mode anyhow.