Fix possible seg fault
Normally, unknown debug client commands are treated as session
lookup requests, allowing us to see the details for an active
session. When running with the -mfw flag the global session table
does not get created, so this would cause a SEGV. I added a check for
the mfw flag before trying to access the session table.


Former-commit-id: 7a873f817d92af08f301384516b15f9e6d37d543