Fixes for NAVL memory usage
For NGFW we disable NAVL tcp and udp timeouts, since we do all of the
session management. Turns out the timeouts aren't used in that
case anyhow so it didn't really matter. But it DOES matter when we
let NAVL handle the session management, in which case using zero
seems like a bad idea. I updated the code to use 60 seconds for both
tcp and udp in MFW mode which should be plenty of time to classify the
first bit of traffic we see before the nft packet limit is triggered.
I also updated our NAVL error counters and debug dump to capture new
error codes and other diagnostic stuff they've added to the library
over the years.

MFW-403 #code-completed